Add inflation-adjusted values to a data frame

This is a wrapper around blscrapeR::bls_api() modeled after blscrapeR::inflation_adjust() that joins a data frame with an inflation adjustment table, then calculates adjusted values. It returns the original data frame with two additional columns for adjustment factors and adjustment values.

adj_inflation(data, value, year, base_year = 2016,
  key = Sys.getenv("BLS_KEY"))

Arguments

data
A data frame containing monetary values by year.

value
Bare column name of monetary values; for safety, has no default.

year
Bare column name of years; for safety, has no default.

base_year
Year on which to base inflation amounts; defaults to 2016.

key
A string giving the BLS API key. Defaults to the value in Sys.getenv("BLS_KEY"), as set by blscrapeR::set_bls_key.

Value

A data frame with two additional columns: adjustment factors, and adjusted values. The adjusted values column is named based on the name supplied as value; e.g. if value = avg_wage, the adjusted column is named adj_avg_wage.

Details

Note: Because adj_inflation wraps around a function that makes API calls, internet access is required.

Examples

  wages <- data.frame(
    year = 2010:2016,
    wage = c(50000, 51000, 52000, 53000, 54000, 55000, 54000)
  )
  adj_inflation(wages, value = wage, year = year, base_year = 2016)

#> REQUEST_SUCCEEDED

#>   year  wage adj_factor adj_wage
#> 1 2010 50000      0.909 55005.50
#> 2 2011 51000      0.937 54429.03
#> 3 2012 52000      0.957 54336.47
#> 4 2013 53000      0.971 54582.90
#> 5 2014 54000      0.986 54766.73
#> 6 2015 55000      0.987 55724.42
#> 7 2016 54000      1.000 54000.00