Quickly add the labels of ACS variables

tidycensus::get_acs returns an ACS table with its variable codes, which can be joined with cwi::acs_vars* to get readable labels. This function is just a quick wrapper around the common task of joining these two data frames.

label_acs(data, year = 2021, survey = "acs5", variable = variable)

Arguments

data

A data frame/tibble.

year

The endyear of ACS data; defaults 2021.

survey

A string: which ACS estimate to use. Defaults to 5-year ("acs5"), but can also be 1-year ("acs1") or 3-year ("acs3"), though both 1-year and 3-year have limited availability.

variable

The bare column name of variable codes; defaults to variable, as returned by tidycensus::get_acs.

Value

A tibble

See also