Fetch a decennial census table with multiple geography levels

Fetch a data table from the decennial census via tidycensus with your choice of geographies at multiple levels. For geographies made of aggregates, i.e. neighborhoods made of tracts or regions made of towns, the returned table will have estimates summed and margins of error calculated for the whole area.

multi_geo_decennial(table, year = 2010, neighborhoods = NULL,
  towns = "all", regions = NULL, counties = "all", state = "09",
  tracts = NULL, sumfile = "sf1", verbose = TRUE)

Arguments

table
A string giving the decennial census table number. These are generally formatted as one or more letters, 3 numbers, and optionally a letter.

year
The year of the census table; currently defaults 2010 (most recent decennial census).

neighborhoods
Temporarily deprecated: A named list of neighborhoods with their 11-digit tract GEOIDs (defaults NULL).

towns
A character vector of towns to include; "all" (default) for all towns optionally filtered by county; or NULL to not fetch town-level table.

regions
A named list of regions with their town names (defaults NULL).

counties
A character vector of counties to include; "all" (default) for all counties in the state; or NULL to not fetch county-level table.

state
A string: either name or two-digit FIPS code of a US state. Required; defaults "09" (Connecticut).

tracts
A character vector of 11-digit FIPS codes of tracts to include, or "all" for all tracts optionally filtered by county. Defaults NULL.

sumfile
A string giving the summary file to pull from. Defaults "sf1"; in some rare cases, "sf3" may be appropriate.

verbose
Logical: whether to print summary of geographies included. Defaults TRUE.

Value

A tibble with GEOID, name, variable code, estimate, moe, geography level, state, and year, as applicable, for the chosen table.

Details

This function essentially calls tidycensus::get_decennial() multiple times, depending on geographic levels chosen, and does minor cleaning, filtering, and aggregation. Note that the underlying tidycensus::get_decennial() requires a Census API key. As is the case with other tidycensus functions, multi_geo_decennial assumes this key is stored as CENSUS_API_KEY in your .Renviron. See tidycensus::census_api_key() for installation.

Be advised that decennial table numbers may change from year to year, so if you're looking at trends, check FactFinder or another source to make sure the tables have the same meaning. Setting verbose = TRUE is helpful for this as well.

See also

Examples

multi_geo_decennial("P001", 2016,
  neighborhoods = list(downtown = c("09009140100", "09009361401", "09009361402"),
    dixwell = "090091416"),
  towns = "all",
  regions = list(inner_ring = c("Hamden", "East Haven", "West Haven")),
  counties = "New Haven County")

#> Error in open.connection(x, "rb"): HTTP error 404.