Get employees counts and total payroll over time

This gets data from the Quarterly Workforce Indicators (QWI) via the Census API. It's an alternative to censusapi that fetches a set of variables (employees and payroll) but makes a somewhat more dynamic API call. The API returns a maximum of 10 years of data; calling this function with more than 10 years will require multiple API calls, which takes a little longer.

qwi_industry(years, industries = naics_codes$industry, counties = NULL,
  state = "09", annual = TRUE, key = Sys.getenv("CENSUS_API_KEY"))

Arguments

years
A numeric vector of one or more years for which to get data

industries
A character vector of NAICS industry codes; default is the 20 sectors plus "All industries" from the dataset naics_codes.

counties
A character vector of county FIPS codes.

state
A string of a state FIPS code; defaults to "09" for Connecticut

annual
Logical, whether to return annual averages (default) or quarterly data.

key
A Census API key. Defaults to the value at "CENSUS_API_KEY" in your .Renviron file, as set by tidycensus::census_api_key().

Value

A data frame / tibble

Details

Note that when looking at data quarterly, the payroll reported will be for that quarter, not the yearly payroll that you may be more accustomed to.

Examples

# \dontrun{
qwi_industry(2012:2017, industries = c("23", "62"), counties = "009")

#> # A tibble: 12 x 6
#>     year industry state county   Emp    Payroll
#>    <dbl> <chr>    <chr> <chr>  <dbl>      <dbl>
#>  1  2012 23       09    009    12860  775481456
#>  2  2012 62       09    009    68712 3281149707
#>  3  2013 23       09    009    13268  813987579
#>  4  2013 62       09    009    69403 3309654225
#>  5  2014 23       09    009    13988  883480249
#>  6  2014 62       09    009    69872 3432956952
#>  7  2015 23       09    009    14629  963915338
#>  8  2015 62       09    009    70324 3453606257
#>  9  2016 23       09    009    15143 1008594663
#> 10  2016 62       09    009    71326 3561217443
#> 11  2017 23       09    009    14851 1010315682
#> 12  2017 62       09    009    70927 3576310162

# }