Add inflation-adjusted values to a data frame

This is modeled after blscrapeR::inflation_adjust that joins a data frame with an inflation adjustment table from the Bureau of Labor Statistics, then calculates adjusted values. It returns the original data frame with two additional columns for adjustment factors and adjustment values.

adj_inflation(
  .data,
  value,
  year,
  base_year = 2019,
  key = Sys.getenv("BLS_KEY")
)

Arguments

.data
A data frame containing monetary values by year.

value
Bare column name of monetary values; for safety, has no default.

year
Bare column name of years; for safety, has no default.

base_year
Year on which to base inflation amounts; defaults to 2019.

key
A string giving the BLS API key. Defaults to the value in Sys.getenv("BLS_KEY").

Value

A data frame with two additional columns: adjustment factors, and adjusted values. The adjusted values column is named based on the name supplied as value; e.g. if value = avg_wage, the adjusted column is named adj_avg_wage.

Details

Note: Because adj_inflation makes API calls, internet access is required.

Examples

if (FALSE) {
  wages <- data.frame(
    fiscal_year = 2010:2016,
    wage = c(50000, 51000, 52000, 53000, 54000, 55000, 54000)
  )
  adj_inflation(wages, value = wage, year = fiscal_year, base_year = 2016)
}