Collapse survey groups and get weighted means

Deprecation notice: Crosstab-related functions have been moved from cwi to the dcws package. The versions here will be removed soon. This is just a quick wrapper for a common, tedious task of collapsing several demographic groups, such as income brackets, into larger groups and taking a weighted mean based on a set of survey weights.

Usage

collapse_n_wt(
  data,
  ...,
  .lvls,
  .group = group,
  .value = value,
  .weight = weight,
  .fill_wts = FALSE,
  .digits = NULL
)

Arguments

data

A data frame, such as returned by dcws::xtab2df() joined with survey weights as returned by dcws::read_weights(). The default column names here match those returned by xtab2df (group, value) and read_weights (weight).

...

Bare column names to use for grouping, including the .group column, such as location, year, category, response, etc–probably everything except values and weights.

.lvls

A named list, where values are character vectors of smaller groups (e.g. c("<$15K", "$15K-$30K")) and names are the groups those will be replaced by (e.g. "<$30K"). This will be split into the arguments to forcats::fct_collapse().

.group

Bare column name of where groups should be found. Default: group

.value

Bare column name of where values should be found. Default: value

.weight

Bare column name of where group weights should be found. Default: weight

.fill_wts

Logical: if TRUE, missing weights will be filled in with 1, i.e. unweighted. This defaults to FALSE, because missing weights is a useful way to find that there's a mismatch between the group labels in the data and those in the weights table, which is very often the case. Therefore, only set this to TRUE if you've already accounted for labeling discrepancies.

.digits

Numeric: if given, weighted means will be rounded to this number of digits. If NULL (the default), values are returned unrounded.

Value

A data frame with summarized values. The .group column will have the collapsed groups, and the .value column will have average values.

See also