Fetch local area unemployment statistics (LAUS) data over time

Fetch monthly LAUS data for a list of locations over a given time period. Requires a BLS API key; see blscrapeR::set_bls_key().

laus_trend(names, startyear, endyear, measures = "all", annual = FALSE,
  key = Sys.getenv("BLS_KEY"))

Arguments

names
A character vector of place names to look up: as of now, this is limited to Connecticut, and its counties and towns.

startyear
Numeric; first year of range

endyear
Numeric; last year of range

measures
A character vector of measures, containing any combination of "unemployment rate", "unemployment", "employment", or "labor force", or "all" (the default) as shorthand for all of the above.

annual
Logical: whether to include annual averages along with monthly data. Defaults FALSE.

key
A string giving the BLS API key. Defaults to the value in Sys.getenv("BLS_KEY"), as set by blscrapeR::set_bls_key.

Value

A data frame / tibble with columns for the measure, area, year, month number and name, value, series ID, and footnotes returned from the API.

Examples

laus_trend(c("New Haven", "Hamden"), 2014, 2017, annual = TRUE)

#> REQUEST_SUCCEEDED

#> # A tibble: 416 x 8
#>    measure_text  area    year period periodName value seriesID   footnotes      
#>    <chr>         <chr>  <dbl> <chr>  <chr>      <dbl> <chr>      <chr>          
#>  1 unemployment… Hamden  2017 M13    Annual       4.3 LAUCS0935… R Data were su…
#>  2 unemployment… Hamden  2017 M12    December     3.5 LAUCS0935… R Data were su…
#>  3 unemployment… Hamden  2017 M11    November     3.7 LAUCS0935… R Data were su…
#>  4 unemployment… Hamden  2017 M10    October      3.9 LAUCS0935… R Data were su…
#>  5 unemployment… Hamden  2017 M09    September    4.1 LAUCS0935… R Data were su…
#>  6 unemployment… Hamden  2017 M08    August       4.6 LAUCS0935… R Data were su…
#>  7 unemployment… Hamden  2017 M07    July         4.8 LAUCS0935… R Data were su…
#>  8 unemployment… Hamden  2017 M06    June         4.7 LAUCS0935… R Data were su…
#>  9 unemployment… Hamden  2017 M05    May          4.4 LAUCS0935… R Data were su…
#> 10 unemployment… Hamden  2017 M04    April        4.1 LAUCS0935… R Data were su…
#> # … with 406 more rows