Fetch an ACS table with multiple geography levels

Fetch a data table from the ACS via tidycensus with your choice of geographies at multiple levels. For geographies made of aggregates, i.e. neighborhoods made of tracts or regions made of towns, the returned table will have estimates summed and margins of error calculated for the whole area.

multi_geo_acs(
  table,
  year = 2017,
  neighborhoods = NULL,
  towns = "all",
  regions = NULL,
  counties = "all",
  state = "09",
  tracts = NULL,
  blockgroups = NULL,
  msa = FALSE,
  us = FALSE,
  new_england = TRUE,
  survey = "acs5",
  verbose = TRUE
)

Arguments

table
A string giving the ACS table number.

year
The year of the ACS table; currently defaults 2017 (most recent available).

neighborhoods
Temporarily deprecated: A named list of neighborhoods with their 11-digit tract GEOIDs (defaults NULL).

towns
A character vector of names of towns to include; "all" (default) for all towns optionally filtered by county; or NULL to not fetch town-level table.

regions
A named list of regions with their town names (defaults NULL).

counties
A character vector of names of counties to include; "all" (default) for all counties in the state; or NULL to not fetch county-level table.

state
A string: either name or two-digit FIPS code of a US state. Required; defaults "09" (Connecticut).

tracts
A character vector of 11-digit FIPS codes of tracts to include, or "all" for all tracts optionally filtered by county. Defaults NULL.

blockgroups
A character vector of 12-digit FIPS codes of block groups to include, or "all" for all block groups optionally filtered by county. Defaults NULL.

msa
Logical: whether to fetch New England states' metropolitan statistical areas. Defaults FALSE.

us
Logical: whether to fetch US-level table. Defaults FALSE.

new_england
Logical: if TRUE (the default), limits metro areas to just New England states.

survey
A string: which ACS estimate to use. Defaults to 5-year ("acs5"), but can also be 1-year ("acs1") or 3-year ("acs3"), though both 1-year and 3-year have limited availability.

verbose
Logical: whether to print summary of geographies included. Defaults TRUE.

Value

A tibble with GEOID, name, variable code, estimate, moe, geography level, state, and year, as applicable, for the chosen ACS table.

Details

This function essentially calls tidycensus::get_acs() multiple times, depending on geographic levels chosen, and does minor cleaning, filtering, and aggregation. Note that the underlying tidycensus::get_acs() requires a Census API key. As is the case with other tidycensus functions, multi_geo_acs assumes this key is stored as CENSUS_API_KEY in your .Renviron. See tidycensus::census_api_key() for installation.

See also

Examples

# \dontrun{
multi_geo_acs("B01003", 2016,
  towns = "all",
  regions = list(inner_ring = c("Hamden", "East Haven", "West Haven")),
  counties = "New Haven County",
  tracts = unique(nhv_tracts$geoid))

#> Table B01003: TOTAL POPULATION

#> Geographies included:
#> Towns: all
#> Regions: inner_ring
#> Counties: New Haven County
#> State: 09

#> # A tibble: 61 x 9
#>    GEOID    NAME        variable  estimate   moe level   state county       year
#>    <chr>    <chr>       <chr>        <dbl> <dbl> <fct>   <chr> <chr>       <dbl>
#>  1 09       Connecticut B01003_0…  3588570    NA 1_state <NA>  <NA>         2016
#>  2 09009    New Haven … B01003_0…   860874    NA 2_coun… 09    <NA>         2016
#>  3 <NA>     inner_ring  B01003_0…   145463    57 3_regi… <NA>  <NA>         2016
#>  4 0900901… Ansonia     B01003_0…    18950    30 4_towns 09    New Haven …  2016
#>  5 0900903… Beacon Fal… B01003_0…     6075    20 4_towns 09    New Haven …  2016
#>  6 0900904… Bethany     B01003_0…     5521    24 4_towns 09    New Haven …  2016
#>  7 0900907… Branford    B01003_0…    28084    28 4_towns 09    New Haven …  2016
#>  8 0900914… Cheshire    B01003_0…    29254    40 4_towns 09    New Haven …  2016
#>  9 0900919… Derby       B01003_0…    12755    25 4_towns 09    New Haven …  2016
#> 10 0900922… East Haven  B01003_0…    29015    24 4_towns 09    New Haven …  2016
#> # … with 51 more rows

# }