City neighborhoods by tract

Datasets of neighborhoods for New Haven, Hartford/West Hartford, Stamford, and Bridgeport. Some tracts cross between more than one neighborhood; use weight column for aggregating values such as populations. Previously this included a block group version for New Haven, which I've removed; I'm also renaming nhv_tracts to new_haven_tracts for consistency.

new_haven_tracts

bridgeport_tracts

stamford_tracts

hartford_tracts

new_haven_tracts19

bridgeport_tracts19

stamford_tracts19

hartford_tracts19

Format

A data frame

town

For hartford_tracts, the name of the town, because both Hartford and West Hartford neighborhoods are included; otherwise, no town variable is needed

name

Neighborhood name

geoid

11-digit FIPS code of the tract

tract

6-digit FIPS code of the tract; same as geoid but missing state & county components.

weight

Share of tract's households in that neighborhood

An object of class tbl_df (inherits from tbl, data.frame) with 37 rows and 3 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 47 rows and 3 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 50 rows and 3 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 64 rows and 4 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 36 rows and 3 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 48 rows and 3 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 39 rows and 3 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 60 rows and 4 columns.

Details

These were updated to 2020 tract definitions. There are still 2019 versions of weight tables with names ending in "19"; for the time being, those will stick around for use with pre-2020 data.