Get employees counts and total payroll over time

This gets data from the Quarterly Workforce Indicators (QWI) via the Census API. It's an alternative to censusapi::getCensus that fetches a set of variables (employees and payroll) but makes a somewhat more dynamic API call. The API returns a maximum of 10 years of data; calling this function with more than 10 years will require multiple API calls, which takes a little longer.

qwi_industry(years, industries = naics_codes$industry, counties = NULL,
  state = "09", annual = TRUE, key = Sys.getenv("CENSUS_API_KEY"))

Arguments

years
A numeric vector of one or more years for which to get data

industries
A character vector of NAICS industry codes; default is the 20 sectors plus "All industries" from the dataset naics_codes.

counties
A character vector of county FIPS codes.

state
A string of a state FIPS code; defaults to "09" for Connecticut

annual
Logical, whether to return annual averages (default) or quarterly data.

key
A Census API key. Defaults to the value at "CENSUS_API_KEY" in your .Renviron file, as set by tidycensus::census_api_key().

Value

A data frame / tibble

Details

Note that when looking at data quarterly, the payroll reported will be for that quarter, not the yearly payroll that you may be more accustomed to.

See also

Examples

qwi_industry(2012:2017, industries = c("23", "62"), counties = "009")

#> # A tibble: 12 x 6
#>     year industry state county   Emp    Payroll
#>    <dbl> <chr>    <chr> <chr>  <dbl>      <dbl>
#>  1  2012 23       09    009    12751  774376073
#>  2  2012 62       09    009    68899 3303327491
#>  3  2013 23       09    009    13131  812396448
#>  4  2013 62       09    009    69866 3352828532
#>  5  2014 23       09    009    13746  874722341
#>  6  2014 62       09    009    69876 3459053684
#>  7  2015 23       09    009    14388  958724428
#>  8  2015 62       09    009    70826 3502470369
#>  9  2016 23       09    009    14941 1010824074
#> 10  2016 62       09    009    71643 3604432327
#> 11  2017 23       09    009    14313  716797663
#> 12  2017 62       09    009    71012 2641709994